Modernize the repo and fix CI + PR flow (#189)
This changeset follows the transfer of the repo to @igor-makarov, the new maintainer.  

Includes:  

* Update the README.md to current maintainer info.
* Fix Travis CI and reduce the outdated test matrix.
* Remove Hound, as it's been providing inconsistent comments, and replace it with Rubocop on Travis.
* Fix most Rubocop linter errors by tweaking the config and fixing remaining offences.
* Configure new Coveralls identity along with SimpleCov for local development.
* Make it easy to `pry` whenever necessary by requiring it in `spec_helper.rb`.

+++ lib/xcake/context/xcodeproj_context.rb
@@ -7,7 +7,7 @@ module Xcake
 
     attr_accessor :project
 
-    def create_object_for(dsl_object)
+    def create_object_for(dsl_object) # rubocop:disable Metrics/CyclomaticComplexity
       case dsl_object
       when BuildPhase
         create_object_for_build_phase(dsl_object)
@@ -28,17 +28,14 @@ module Xcake
       @project.new(build_phase.build_phase_type)
     end
 
-    def create_object_for_build_rule(build_rule)
+    def create_object_for_build_rule(_build_rule)
       @project.new(Xcodeproj::Project::Object::PBXBuildRule)
     end
 
     def create_object_for_project(project)
-      
       project_path = "./#{project.name}.xcodeproj"
-      
-      if File.exist?(project_path)
-        FileUtils.remove_dir(project_path)
-      end
+
+      FileUtils.remove_dir(project_path) if File.exist?(project_path)
 
       @project = Xcode::Project.new(project_path, true)
       @project.setup_for_xcake
@@ -53,7 +50,7 @@ module Xcake
       @project.new_configuration(configuration)
     end
 
-    def create_object_for_scheme(scheme)
+    def create_object_for_scheme(_scheme)
       Xcode::Scheme.new
     end
